Clean the DAO class and move logical methods to the storage service
