+++ exo-gadget-pack/gadget-pack-services/src/main/java/org/exoplatform/platform/gadget/services/LoginHistory/jpa/entity/LoginHistoryEntity.java
@@ -10,48 +10,50 @@ import java.sql.Timestamp;
 @ExoEntity
 @Table(name = "login_history")
 @NamedQueries({
+        @NamedQuery(name = "loginHistory.getLastLoginHistory",query = "SELECT * FROM LoginHistoryEntity l WHERE l.userID = :userId ORDER BY ID DESC LIMIT 1"),
         @NamedQuery(name = "loginHistory.getUserLastLoginID",query = "SELECT MAX(l.ID) FROM LoginHistoryEntity l WHERE l.userID = :userId"),
         @NamedQuery(name = "loginHistory.getBeforeLastLoginID",query = "SELECT MAX(l.ID) FROM LoginHistoryEntity l WHERE l.userID = :userId AND l.ID < :id"),
         @NamedQuery(name = "loginHistory.getUserLoginHistory",query = "SELECT * FROM LoginHistoryEntity l WHERE l.userID = :userId AND l.loginDate BETWEEN :from AND :to"),
         @NamedQuery(name = "loginHistory.getLastLogins",query = "SELECT * FROM LoginHistoryEntity l WHERE l.userID = :userId AND l.loginDate < :today LIMIT :limit"),
-        // @NamedQuery(name = "loginHistory.getLoginPerDayInRange",query = "SELECT * FROM LoginHistoryEntity l WHERE l.userID = :userId AND l.loginDate BETWEEN :from AND :to ORDER BY l.loginDate ASC"),
+        @NamedQuery(name = "loginHistory.addLoginHistory",query = "INSERT INTO LoginHistoryEntity l (l.userID) VALUES (:userId)"),
         @NamedQuery(name = "loginHistory.getLoginCountPerDay",query = "SELECT COUNT (*) FROM LoginHistoryEntity l WHERE l.userID = :userId AND l.loginDate BETWEEN :from AND :to"),
-        @NamedQuery(name = "loginHistory.getLastUsersLogin",query = "SELECT l.userID FROM LoginHistoryEntity l WHERE l.loginDate >= :from ORDER BY l.ID DESC")
+        @NamedQuery(name = "loginHistory.getLastUsersLogin",query = "SELECT l.userID FROM LoginHistoryEntity l WHERE l.loginDate >= :from ORDER BY l.ID DESC"),
+        @NamedQuery(name = "loginHistory.getActiveUsersId",query = "SELECT DISTINCT l.userId FROM LoginHistoryEntity l where l.loginDate < :from")
 })
 public class LoginHistoryEntity {
     @Id
     @GeneratedValue
     @Column(name = "ID")
-    private long ID;
+    private long id;
 
     @Column(name = "USER_ID")
-    private String userID;
+    private String userId;
 
-    @Column(name = "DATE")
+    @Column(name = "LOGIN_DATE")
     private Timestamp loginDate;
 
     public LoginHistoryEntity() {
     }
 
     public LoginHistoryEntity(String userID) {
-        this.userID = userID;
+        this.userId = userID;
     }
 
     public LoginHistoryEntity(String userID, Timestamp loginDate) {
-        this.userID = userID;
+        this.userId = userID;
         this.loginDate = loginDate;
     }
 
     public long getID() {
-        return ID;
+        return id;
     }
 
     public String getUserID() {
-        return userID;
+        return userId;
     }
 
     public void setUserID(String userID) {
-        this.userID = userID;
+        this.userId = userID;
     }
 
     public Timestamp getLoginDate() {
