+++ exo-gadget-pack/gadget-pack-services/src/main/java/org/exoplatform/platform/gadget/services/LoginHistory/storage/JPALoginHistoryStorageImpl.java
@@ -1,20 +1,114 @@
 package org.exoplatform.platform.gadget.services.LoginHistory.storage;
 
 import org.exoplatform.commons.api.persistence.ExoTransactional;
+import org.exoplatform.commons.persistence.impl.GenericDAOJPAImpl;
+import org.exoplatform.container.ExoContainerContext;
 import org.exoplatform.platform.gadget.services.LoginHistory.LastLoginBean;
 import org.exoplatform.platform.gadget.services.LoginHistory.LoginCounterBean;
 import org.exoplatform.platform.gadget.services.LoginHistory.LoginHistoryBean;
 import org.exoplatform.platform.gadget.services.LoginHistory.jpa.dao.LoginHistoryDAO;
+import org.exoplatform.platform.gadget.services.LoginHistory.jpa.entity.LoginHistoryEntity;
+import org.exoplatform.services.log.ExoLogger;
+import org.exoplatform.services.log.Log;
+import org.exoplatform.services.organization.OrganizationService;
 
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.sql.Timestamp;
+import java.util.*;
 
-public class JPALoginHistoryStorageImpl implements LoginHistoryStorage {
+public class JPALoginHistoryStorageImpl extends GenericDAOJPAImpl implements LoginHistoryStorage {
+    private static final Log LOG = ExoLogger.getLogger(LoginHistoryDAO.class);
+    private static long DAY_IN_MILLISEC = 86400000;
     private LoginHistoryDAO loginHistoryDAO;
 
     public JPALoginHistoryStorageImpl(LoginHistoryDAO loginHistoryDAO) {
         this.loginHistoryDAO = loginHistoryDAO;
+    }
+
+    private String getUserFullName(String userId) {
+        try {
+            OrganizationService service = (OrganizationService) ExoContainerContext.getCurrentContainer()
+                    .getComponentInstanceOfType(OrganizationService.class);
+            return service.getUserHandler().findUserByName(userId).getFullName();
+        } catch (Exception e) {
+            return userId;
+        }
+    }
+
+    private static long nextMonday(long date) {
+        Calendar now = Calendar.getInstance();
+        now.setTimeInMillis(date);
+
+        int weekday = now.get(Calendar.DAY_OF_WEEK);
+        // calculate how much to add
+        // the 2 is the difference between Saturday and Monday
+        int days = weekday == Calendar.SUNDAY ? 1 : Calendar.SATURDAY - weekday + 2;
+        now.add(Calendar.DAY_OF_YEAR, days);
+        return now.getTimeInMillis();
+    }
+
+    public List<LoginCounterBean> getLoginCountPerDaysInRange(String userId, long fromDate, long toDate) throws Exception {
+        try {
+            // set the fromDate to 00:00
+            Calendar from = Calendar.getInstance();
+            Calendar to = Calendar.getInstance();
+            from.setTimeInMillis(fromDate);
+            to.setTimeInMillis(toDate);
+            from.set(Calendar.MILLISECOND,0);
+            from.set(Calendar.SECOND,0);
+            from.set(Calendar.MINUTE,0);
+            from.set(Calendar.HOUR,0);
+
+            //set the toDate to 23:59
+            to.set(Calendar.MILLISECOND,999);
+            to.set(Calendar.SECOND,59);
+            to.set(Calendar.MINUTE,59);
+            to.set(Calendar.HOUR,23);
+
+            // instantiate first day, next day and lastDay variables
+            long firstDay = from.getTimeInMillis();
+            Long nextDay = 0L;
+            Long lastDay = to.getTimeInMillis();
+
+            List<LoginCounterBean> counterBeanList = new ArrayList<LoginCounterBean>();
+            LoginCounterBean loginCountPerDay = new LoginCounterBean();
+
+            // returns the user's login count for each day and add it to a list: loginCount/day
+            while (firstDay <= lastDay) {
+                Calendar cal = Calendar.getInstance();
+                cal.setTimeInMillis(firstDay);
+                cal.set(Calendar.DAY_OF_YEAR, cal.get(Calendar.DAY_OF_YEAR) +1);
+                nextDay = cal.getTimeInMillis();
+
+                Calendar cal2 = Calendar.getInstance();
+                cal2.setTimeInMillis(firstDay);
+                cal2.set(Calendar.MILLISECOND,999);
+                cal2.set(Calendar.SECOND,59);
+                cal2.set(Calendar.MINUTE,59);
+                cal2.set(Calendar.HOUR_OF_DAY,23);
+                Long endOfDay = cal2.getTimeInMillis();
+
+                Long count = (Long) loginHistoryDAO.getLoginCountPerDay(userId,firstDay,endOfDay);
+                loginCountPerDay.setLoginCount(count);
+                loginCountPerDay.setLoginDate(firstDay);
+                counterBeanList.add(loginCountPerDay);
+
+                firstDay = nextDay;
+            }
+            return counterBeanList;
+        } catch (Exception e) {
+            throw e;
+        }
+    }
+
+    private int getLoginCountInDateRange(String userId, long fromDate, long toDate) throws Exception {
+        List<LoginCounterBean> loginCounts = getLoginCountPerDaysInRange(userId, fromDate, toDate);
+
+        int sum = 0;
+        Iterator<LoginCounterBean> iter = loginCounts.iterator();
+        while (iter.hasNext()) {
+            sum += iter.next().getLoginCount();
+        }
+        return sum;
     }
 
     @Override
@@ -24,48 +118,179 @@ public class JPALoginHistoryStorageImpl implements LoginHistoryStorage {
 
     @Override
     public List<LastLoginBean> getLastLogins(int numLogins, String userIdFilter) throws Exception {
-        return null;
+            String userName = getUserFullName(userIdFilter);
+            Calendar calendar = Calendar.getInstance();
+            calendar.set(Calendar.HOUR_OF_DAY,0);
+            calendar.set(Calendar.MINUTE,0);
+            calendar.set(Calendar.SECOND,0);
+            calendar.set(Calendar.MILLISECOND,0);
+            Timestamp today = new Timestamp(calendar.getTimeInMillis());
+
+            List<LoginHistoryEntity> loginHistoryEntityList = loginHistoryDAO.getLastLogins(numLogins,userIdFilter);
+            LastLoginBean lastLoginBean = new LastLoginBean();
+            List<LastLoginBean> lastLoginBeanList = new ArrayList<LastLoginBean>();
+
+            for (int i=0; i<loginHistoryEntityList.size(); i++) {
+                lastLoginBean.setUserId(loginHistoryEntityList.get(i).getUserID());
+                lastLoginBean.setUserName(userName);
+                lastLoginBean.setLastLogin(getLastLogin(userIdFilter));
+                lastLoginBean.setBeforeLastLogin(getBeforeLastLogin(userIdFilter));
+                lastLoginBeanList.add(lastLoginBean);
+            }
+
+            return lastLoginBeanList;
     }
 
-    @Override
     @ExoTransactional
-    public void addLoginHistoryEntry(String userId, long loginTime) throws Exception {
-        loginHistoryDAO.addLoginHistoryEntry(userId,loginTime);
+    public void addLoginHistoryEntry(String userId, long loginTime) throws  Exception {
+        try {
+            LoginHistoryEntity loginHistoryEntity = new LoginHistoryEntity(userId);
+            create(loginHistoryEntity); //the create method will return the entity which we'll ignore.
+        } catch (Exception e) {
+            throw e;
+        }
     }
 
     @Override
     public List<LoginHistoryBean> getLoginHistory(String userId, long fromTime, long toTime) throws Exception {
-        return null;
+        Timestamp from = new Timestamp(fromTime);
+        Timestamp to = new Timestamp(toTime);
+        List<LoginHistoryEntity> loginHistoryEntityList = getEntityManager().createNamedQuery("loginHistory.getUserLoginHistory")
+                .setParameter("userId",userId)
+                .setParameter("from",from)
+                .setParameter("to",to).getResultList();
+        List<LoginHistoryBean> loginHistoryBeanList = new ArrayList<LoginHistoryBean>();
+        LoginHistoryBean loginHistoryBean = new LoginHistoryBean();
+        String userName = getUserFullName(userId);
+
+        try {
+            for (int i=0; i<loginHistoryEntityList.size(); i++) {
+                loginHistoryBean.setUserId(loginHistoryEntityList.get(i).getUserID());
+                loginHistoryBean.setUserName(userName);
+                loginHistoryBean.setLoginTime(loginHistoryEntityList.get(i).getLoginDate().getTime());
+                loginHistoryBeanList.add(loginHistoryBean);
+            }
+            return loginHistoryBeanList;
+        } catch (Exception e) {
+            throw e;
+        }
     }
 
     @Override
     public Set<String> getLastUsersLogin(long fromTime) throws Exception {
-        return null;
+        return loginHistoryDAO.getLastUsersLogin(fromTime);
     }
 
-    @Override
-    public boolean isActiveUser(String userId, int days) {
-        return false;
+    public boolean isActiveUser(String userId, int days) throws Exception {
+        Long beforeLastLogin = getBeforeLastLogin(userId);
+        // return true if it's the first login of user
+        if (beforeLastLogin == 0) return true;
+        //
+        Calendar calendar = Calendar.getInstance();
+        calendar.set(Calendar.DAY_OF_MONTH,calendar.get(Calendar.DAY_OF_MONTH) - days);
+        long limitTime = calendar.getTimeInMillis();
+        return beforeLastLogin >= limitTime;
     }
 
     @Override
     public Map<String, Integer> getActiveUsers(long fromTime) {
-        return null;
+        Map<String, Integer> users = new LinkedHashMap<String, Integer>();
+        List<String> list = loginHistoryDAO.getActiveUsersId(fromTime);
+        Timestamp timestamp = new Timestamp(System.currentTimeMillis());
+        Long toTime = timestamp.getTime();
+        for (int i=0; i<list.size(); i++) {
+            users.put(list.get(i), Math.toIntExact(loginHistoryDAO.getLoginCountPerDay(list.get(i), fromTime, toTime)));
+        }
+        return users;
     }
 
     @Override
     public List<LoginCounterBean> getLoginCountPerDaysInWeek(String userId, long week) throws Exception {
-        return null;
+        List<LoginCounterBean> list = new ArrayList<LoginCounterBean>();
+        List<Long> days = new ArrayList<Long>();
+
+        long now = System.currentTimeMillis();
+        long nextWeek = nextMonday(week);
+
+        long day = week;
+        LoginCounterBean loginCountPerDay;
+
+        do {
+            loginCountPerDay = new LoginCounterBean();
+            loginCountPerDay.setLoginDate(day);
+            loginCountPerDay.setLoginCount(day > now ? -1 : 0);
+
+            list.add(loginCountPerDay);
+            days.add(day);
+
+            day += DAY_IN_MILLISEC;
+        } while (day < nextWeek);
+
+        List<LoginCounterBean> counters = getLoginCountPerDaysInRange(userId, week, nextMonday(week) - DAY_IN_MILLISEC);
+
+        Iterator<LoginCounterBean> iter = counters.iterator();
+        while (iter.hasNext()) {
+            loginCountPerDay = iter.next();
+            list.set(days.indexOf(loginCountPerDay.getLoginDate()), loginCountPerDay);
+        }
+
+        return list;
     }
 
     @Override
     public List<LoginCounterBean> getLoginCountPerWeeksInMonths(String userId, long fromMonth, int numOfMonths) throws Exception {
-        return null;
+        Calendar cal = Calendar.getInstance();
+        long now = cal.getTime().getTime();
+
+        cal.setTimeInMillis(fromMonth);
+        cal.add(Calendar.MONTH, numOfMonths);
+        long toMonth = cal.getTimeInMillis();
+
+        long fromDate, toDate = fromMonth;
+        List<LoginCounterBean> list = new ArrayList<LoginCounterBean>();
+
+        do {
+            fromDate = toDate;
+            toDate = nextMonday(toDate);
+            if (toDate > toMonth) toDate = toMonth;
+
+            LoginCounterBean loginCountPerWeek = new LoginCounterBean();
+            loginCountPerWeek.setLoginDate(fromDate);
+            loginCountPerWeek.setLoginCount(fromDate > now ? -1 : getLoginCountInDateRange(userId, fromDate, toDate - DAY_IN_MILLISEC));
+
+            list.add(loginCountPerWeek);
+        } while (toDate < toMonth);
+
+        return list;
     }
 
     @Override
     public List<LoginCounterBean> getLoginCountPerMonthsInYear(String userId, long year) throws Exception {
-        return null;
+        Calendar cal = Calendar.getInstance();
+        long now = cal.getTime().getTime();
+
+        cal.setTimeInMillis(year);
+        cal.add(Calendar.YEAR, 1);
+        long nextYear = cal.getTimeInMillis();
+
+        long fromDate, toDate = year;
+        List<LoginCounterBean> list = new ArrayList<LoginCounterBean>();
+
+        do {
+            fromDate = toDate;
+            cal.setTimeInMillis(toDate);
+            cal.add(Calendar.MONTH, 1);
+            toDate = cal.getTimeInMillis();
+            if (toDate > nextYear) toDate = nextYear;
+
+            LoginCounterBean loginCountPerWeek = new LoginCounterBean();
+            loginCountPerWeek.setLoginDate(fromDate);
+            loginCountPerWeek.setLoginCount(fromDate > now ? -1 : getLoginCountInDateRange(userId, fromDate, toDate - DAY_IN_MILLISEC));
+
+            list.add(loginCountPerWeek);
+        } while (toDate < nextYear);
+
+        return list;
     }
 
     @Override
