+++ exo-gadget-pack/gadget-pack-services/src/main/java/org/exoplatform/platform/gadget/services/LoginHistory/storage/LoginHistoryStorage.java
@@ -38,7 +38,7 @@ public interface LoginHistoryStorage {
 
     public Set<String> getLastUsersLogin(long fromTime) throws Exception;
 
-    public boolean isActiveUser(String userId, int days);
+    public boolean isActiveUser(String userId, int days) throws Exception;
 
     public Map<String, Integer> getActiveUsers(long fromTime);
